Feedback service taasloodud ja töötab

+++ backend/src/main/java/ee/nutrionista/persistence/feedback/FeedbackRepository.java
@@ -2,5 +2,5 @@ package ee.nutrionista.persistence.feedback;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FeedbackRepository extends JpaRepository<Feedback, Long> {
+public interface FeedbackRepository extends JpaRepository<Feedback, Integer> {
 }
